+++ src/main/kotlin/com/wafflestudio/team03server/config/AwsConfig.kt
@@ -4,7 +4,7 @@ import com.amazonaws.auth.AWSStaticCredentialsProvider
 import com.amazonaws.auth.BasicAWSCredentials
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.AmazonS3ClientBuilder
-import com.wafflestudio.team03server.core.user.service.S3Properties
+import com.wafflestudio.team03server.properties.S3Properties
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
