+++ src/main/kotlin/com/wafflestudio/team03server/utils/EmailService.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.team03server.core.user.service
+package com.wafflestudio.team03server.utils
 
 import org.springframework.mail.javamail.JavaMailSender
 import org.springframework.mail.javamail.MimeMessageHelper
@@ -21,11 +21,11 @@ class EmailService(
         mailSender.send(message)
     }
 
-    fun sendVerificationEmail(email: String, verificationToken: String) {
+    fun sendVerificationEmail(email: String, code: String) {
         sendEmail(
             email,
-            "회원가입 인증 이메일",
-            "http://localhost:8080/auth/verifyEmail?token=$verificationToken",
+            "회원가입 인증 코드",
+            "인증 코드: $code",
         )
     }
 }
